minor warning workaround
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/EBImage@22735 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/colors.c
@@ -257,7 +257,7 @@ image1D2CHAR (Image * image) {
     PROTECT (res = allocVector(STRSXP, nvals) );
     nprotect++;
 
-    for ( i = 0; i < image->columns; i++ ) {
+    for ( i = 0; i < (int)image->columns; i++ ) {
         pp = GetOnePixel (image, i, 0);
         strcpy (pixelStr, "#");
         sprintf (component, "%02X", ScaleQuantumToChar(pp.red) );
